Update create_hash renamed to game
