+++ test/game_test.rb
@@ -1,10 +1,10 @@
 require "minitest/autorun"
 require "minitest/pride"
-require "./lib/create_hash"
+require "./lib/game"
 require 'csv'
 require 'pry'
 
-class CreateHashTest < MiniTest::Test
+class GameTest < MiniTest::Test
 
   def setup
     game_path = './data/dummy_file_games.csv'
@@ -19,24 +19,24 @@ class CreateHashTest < MiniTest::Test
   end
 
   def test_it_exists
-    new_hash = CreateHash.new
+    game = Game.new
 
-    assert_instance_of CreateHash, new_hash
+    assert_instance_of Game, game
   end
 
   def test_create_hash_keys
     array_dummy = CSV.read('./data/dummy_file_games.csv')
-    new_hash = CreateHash.new()
-    new_hash.create_stat_hash_keys(array_dummy)
+    game = Game.new()
+    game.create_stat_hash_keys(array_dummy)
 
-    assert_equal 10, new_hash.stat_hash.keys.size
+    assert_equal 10, game.stat_hash.keys.size
   end
 
   def test_create_stat_hash_values
     array_dummy = CSV.read('./data/dummy_file_games.csv')
-    new_hash = CreateHash.new()
-    new_hash.create_stat_hash_keys(array_dummy)
+    game = Game.new
+    game.create_stat_hash_keys(array_dummy)
 
-    assert_equal 19, new_hash.stat_hash["game_id"].size
+    assert_equal 19, game.stat_hash["game_id"].size
   end
 end
